Make esbuild optional
esbuild is a large dependency, and pinning to a specific version makes it less likely that it can be shared across other dependencies. To make i18next-parser lighter:

- Dynamically import esbuild if needed. Change esbuild to an optional peer dependency.
- Update the version range.
- Try loading .ts files using Node.js support for stripping TypeScript types (enabled by default in current Node.js versions), to potentially remove having to import ESBuild at all.
- Allow loading .mts as well as .ts.

This is a potentially breaking change: if a someone uses a .ts i18next-parser configuration, doesn't have type stripping enabled (or has a configuration file that's incompatible with Node.js's type stripping), and doesn't otherwise have esbuild installed, then they'll have to install it.